consolidating PageDetails code
